+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.0.3] - 2024-02-01
+
+### Added
+- Customized & `export`ed copy of sanitizer config
+- Add functions to append or prepend or replace (`DocumentFragment`, `Node`, `CSSStyleSheet`, etc)
+- Add `darkCSS` & `lightCSS`
+
+### Changed
+- Update various formatting methods
+- Add handling for `<svg>` missing `xmlns`
+
 ## [v0.0.2] - 2024-01-26
 
 ### Added